requirements.txt
opencv-python==4.6.0
pafy==0.5.5
pillow==9.1.1